Fix: Modularizacion del codigo
- Se agrega la carpeta services donde se almacena el manejador de
  conexiones al lobby
- Test se ubica en su archivo correspondiente

 Cambios a ser confirmados:
	modificados:     actualizar_lobby.py
	modificados:     main.py
	nuevos archivos: services/connection_manager.py
	nuevos archivos: test_actualizar_lobby.py
