+++ main.py
@@ -1,5 +1,5 @@
 from os import getenv
-from fastapi import FastAPI, Response, Request, Depends
+from fastapi import FastAPI, Response, Request, Depends, WebSocket, WebSocketDisconnect
 from sqlalchemy.orm import Session
 
 from database import Database
